add crud

+++ lab2/NavigatorService/src/main/java/ru/ifmo/navigator/service/impl/StarshipServiceImpl.java
@@ -55,7 +55,7 @@ public class StarshipServiceImpl implements StarshipService {
         starShip.setName(request.getName());
         starShip.setFleet(request.getFleet());
         starShip.setCoordinateX(request.getCoordinates().getX());
-        starShip.setCoordinateY(request.getCoordinates().getY());
+        starShip.setCoordinateY(request.getCoordinates().getY().longValue());
 
         starshipRepository.save(starShip);
     }
